requirements.txt
langchain
python-dotenv
ipykernel
langchain_community
streamlit